+++ app/components/github_integration/webhooks/commits.py
@@ -20,7 +20,6 @@ from app.common.linker import (
 )
 from app.components.github_integration.entities.resolution import resolve_repo_signature
 from app.components.github_integration.models import GitHubUser
-from app.components.github_integration.webhooks import monalisten_client
 from app.components.github_integration.webhooks.core import (
     EmbedContent,
     Footer,
@@ -260,7 +259,3 @@ class Commits(commands.Cog):
             view_type=CommitActions,
             view_timeout=60,
         )
-
-
-async def setup(bot: GhosttyBot) -> None:
-    await bot.add_cog(Commits(bot, monalisten_client))
